// Jobs.Frontend/src/app/core/api/model/updateJobPostingRequest.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobCategory } from './jobCategory';


/**
 * Request DTO for updating a job posting.
 */
export interface UpdateJobPostingRequest { 
    title: string;
    description?: string | null;
    category?: JobCategory;
    minHoursPerWeek?: number | null;
    maxHoursPerWeek?: number | null;
    allowsRemoteWork?: boolean | null;
    contactEmail?: string | null;
    jobUrl?: string | null;
}
export namespace UpdateJobPostingRequest {
}